Complete Source

+++ reactjs/src/Components/table.js
@@ -7,10 +7,17 @@ import App from "../App";
 import LEDStatus from "./ledStatus";
 import { setGlobalValue } from "./Constant.js";
 import  { useEffect } from "react";
+ //import 'bootstrap/dist/css/bootstrap.min.css';
+ import Cookies from 'js-cookie';
 
 
 const handleDeleteDevice = (index) => {
-  deleteDevice(index);
+ 
+  const shouldDelete = window.confirm(`Bạn có chắc muốn xoá thiết bị không?`);
+  if (shouldDelete) {
+    deleteDevice(index);
+  }
+
 };
 
 const deleteDevice = (index) => {
@@ -26,34 +33,34 @@ const deleteDevice = (index) => {
     });
 }
 
-
-
-
 const Table = ({ data }) => {
   const [columns, setColumns] = useState([
     {
-      header: "deviceName",
+      header: "Tên thiết bị",
       value: "deviceName",
     },
     {
-      header: "topicName",
+      header: "TopicMQTT",
       value: "topicName",
     },
-    {
-      header: "Location",
-      value: "location",
-    },
+    // {
+    //   header: "Location",
+    //   value: "location",
+    // },
   ]);
   // const cpuUsage = 70;
   // const ramUsage = 45;
   
   
   const handleClick = (value) => {
-    setGlobalValue(value);  }; // Thay đổi giá trị của biến global
+    // setGlobalValue(value); 
+    Cookies.set('globalValue', value);
+
+  }; // Thay đổi giá trị của biến global
   return (
     <div>
       {data.map((row, index) => (
-        <div key={index} style={{ border: "1px solid #ccc", padding: "10px", margin: "10px" }}>
+        <div key={index} style={{ border: "2px solid #ccc", padding: "10px", margin: "10px",backgroundColor: '#f8f8f8' }}>
           {columns.map((column) => (
             <p key={column.value}>
               <strong>{column.header}:</strong> {row[column.value]}
@@ -61,14 +68,14 @@ const Table = ({ data }) => {
             </p>
           ))}
 
-          <Link to={`/edit/${row.deviceName}`} style={{ marginRight: "20px" }} onClick={() => handleClick(data[index][columns[0].value])}  >
+          <Link to={`/edit/${row.deviceName}`}  onClick={() => handleClick(data[index][columns[0].value])} class="btn btn-outline-success" style={{ marginRight:'20px' }}  >
             Chi tiết 
           </Link>
           {/* <CpuRamChart cpu={cpuUsage} ram={ramUsage} /> */}
           {/* <Routes>
           <Route path={`/edit/iot2050`} index element={<CpuRamChart cpu={cpuUsage} ram={ramUsage} />} />
           </Routes> */}
-          <button onClick={() => handleDeleteDevice(data[index][columns[0].value])}>
+          <button class="btn btn-outline-danger"  onClick={() => handleDeleteDevice(data[index][columns[0].value])} >
             Xoá thiết bị
           </button>
         </div>
